Moved bearer token into its own class
