+++ search-tweets-java-core/src/main/java/dev/twitter/api/v2/model/token/BearerToken.java
@@ -1,0 +1,8 @@
+package dev.twitter.api.v2.model.token;
+
+import lombok.Data;
+
+@Data
+public class BearerToken {
+  private String bearerToken;
+}
